requirements.txt
fastapi==0.118.2
uvicorn==0.37.0
httpx==0.28.1
websockets==15.0.1
aiohttp==3.13.0
anyio==4.11.0
cartesia==2.0.9
supabase==2.22.0
livekit-agents[cartesia]~=1.2  # Bundles livekit, livekit-api, and cartesia plugin
python-dotenv==1.1.1
pydantic==2.12.0
aiofiles==24.1.0
numpy==2.0.1
